restructure version module
